Rename global exercise hints file
This was a very difficult file to name correctly.

We tried SETUP.md for a long time. But that's not really what it is, and it was confusing.
Then we tried TRACK_HINTS.md, which was better, but still not very cleary.

Also we weren't sure where to put it. At first, root. Then in the exercises/ directory (but
it's not an exercise, so that wasn't the best location for it, in retrospect).

Finally we concluded that

a) it should be in the docs/ directory, because it's user-facing documentation, and
b) it should be named EXERCISE_README_INSERT.md because that's what it is.

See https://github.com/exercism/meta/issues/5 for context.