Update the issue template with questions about logos
This encodes some of the knowledge that came out of the great logo research project for nextercism.

Some languages have both official and unofficial logos (e.g. PHP has an official logo, and also uses an elephant as
an unofficial logo, whereas Elm has an official logo consisting of a tangram, and unofficially people reorganize
the tangram in creative ways to create custom logos. Go has an official logo and mascot, and many people create variations
on that mascot for their own purposes. Java has very, very strict rules about where you can use the logo, and we had
to create an 'artists impression' of the logo rather than using it directly.