Tweak bootstrap script
The script blows up and exits if we are unable to add a user to the maintainers
team for any reason. That might be a typo, or that the folks who opened the
request for the new track listed an org rather than a user.

This just moves on if that happens.

We will see the error on the command line, and can then do stuff manually
to fix it, if necessary.

This also moves the maintainers-admin team bit up to before the usernames.